+++ code/test/sixsq/nuvla/server/resources/event_utils_test.clj
@@ -1,8 +1,8 @@
 (ns sixsq.nuvla.server.resources.event-utils-test
   (:require [clojure.test :refer [deftest is testing use-fixtures]]
+            [sixsq.nuvla.server.middleware.authn-info :refer [authn-info-header]]
             [sixsq.nuvla.server.resources.common.utils :as u]
             [sixsq.nuvla.server.resources.event.utils :as t]
-            [sixsq.nuvla.server.middleware.authn-info :refer [authn-info-header]]
             [sixsq.nuvla.server.resources.lifecycle-test-utils :as ltu]
             [sixsq.nuvla.server.util.time :as time]))