+++ code/test/sixsq/nuvla/server/resources/ts_nuvlaedge_telemetry_lifecycle_test.clj
@@ -7,8 +7,8 @@
     [sixsq.nuvla.server.app.params :as p]
     [sixsq.nuvla.server.middleware.authn-info :refer [authn-info-header]]
     [sixsq.nuvla.server.resources.common.utils :as u]
-    [sixsq.nuvla.server.resources.ts-nuvlaedge-telemetry :as t]
     [sixsq.nuvla.server.resources.lifecycle-test-utils :as ltu]
+    [sixsq.nuvla.server.resources.ts-nuvlaedge-telemetry :as t]
     [sixsq.nuvla.server.util.time :as time]))
 
 (use-fixtures :each ltu/with-test-server-fixture)